+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Fix splat on Windows not using forward slashes on generated paths.
 - Setup CI to be run on Windows and Macos too.
 - Fix `asmtu` segments not writing the other corresponding sections into the generated assembly file.
+- New yaml option: `keep_complete_c_asm_splits_for_objdiff_compatibility`
+  - Emit a full `.s` file for each `c`/`cpp` segment besides the generated `nonmatchings` individual functions.
+  - Can be used to generate "base" or "expected" objects for asm diffing.
 
 ### 0.32.3